Check all route for non connected people and correct route autorization

+++ src/Controller/CommentController.php
@@ -9,13 +9,18 @@ use Symfony\Component\Routing\Annotation\Route;
 
 class CommentController extends AbstractController
 {
-    #[Route('/{id}/delete', name: 'delete_comment')]
+    #[Route('/delete/{id}', name: 'delete_comment')]
     public function delete(Comment $comment): Response{
-        $entityManager = $this->getDoctrine()->getManager();
-        $entityManager->remove($comment);
-        $entityManager->flush();
-        return $this->redirectToRoute('recipe_show', [
-            'id' => $comment->getRecette()->getId(),
-        ]);
+        $user = $this->getUser();
+        if(in_array("ROLE_ADMIN", $user->getRoles()) XOR $user->getId() == $comment->getUser()->getId()){
+            $entityManager = $this->getDoctrine()->getManager();
+            $entityManager->remove($comment);
+            $entityManager->flush();
+            return $this->redirectToRoute('recipe_show', [
+                'id' => $comment->getRecette()->getId(),
+            ]);
+        }else{
+            return $this->redirectToRoute('login');
+        }
     }
 }
